Change addScore to module
